Add command subfolder to cmd/legacy
