--- cmd/exitcodes.go
@@ -1,3 +1,0 @@
-package cmd
-
-const errCodeDefault = 1
