--- cmd/legacy/exitcodes.go
@@ -1,7 +1,0 @@
-package legacy
-
-const (
-	successCode            = 0
-	errCodeDefault         = 1
-	errCodeConfigFileParse = 103
-)
