+++ cmd/legacy/viper_test.go
@@ -78,67 +78,3 @@ func TestConfigFilePath(t *testing.T) {
 		})
 	}
 }
-
-func TestConfig_LoadConfigReadParams(t *testing.T) {
-	v := viper.New()
-	v.Set("config-section", "settings")
-	v.Set("config-read", "api_key")
-
-	params, err := legacy.LoadConfigReadParams(v)
-	require.NoError(t, err)
-
-	assert.Equal(t, legacy.ConfigReadParams{
-		Key:     "api_key",
-		Section: "settings",
-	}, params)
-}
-
-func TestConfig_LoadConfigReadParamsErr(t *testing.T) {
-	tests := map[string]struct {
-		Key     string
-		Section string
-	}{
-		"section_missing": {
-			Key: "api_key",
-		},
-		"key_missing": {
-			Section: "settings",
-		},
-		"all_missing": {},
-	}
-
-	for name, test := range tests {
-		t.Run(name, func(t *testing.T) {
-			v := viper.New()
-			v.Set("config-section", test.Section)
-			v.Set("config-read", test.Key)
-
-			_, err := legacy.LoadConfigReadParams(v)
-
-			var cfrerr legacy.ErrConfigFileRead
-			assert.True(t, errors.As(err, &cfrerr))
-		})
-	}
-}
-
-func TestConfig_RunConfigRead(t *testing.T) {
-	v := viper.New()
-	v.Set("config-section", "settings")
-	v.Set("config-read", "api_key")
-	v.Set("settings.api_key", "b9485572-74bf-419a-916b-22056ca3a24c")
-
-	err := legacy.RunConfigRead(v)
-	require.NoError(t, err)
-}
-
-func TestConfig_RunConfigReadErr(t *testing.T) {
-	v := viper.New()
-	v.Set("config-section", "settings")
-	v.Set("config-read", "unset_key")
-
-	err := legacy.RunConfigRead(v)
-
-	var cfrerr legacy.ErrConfigFileRead
-
-	assert.True(t, errors.As(err, &cfrerr))
-}
