+++ pkg/config/config.go
@@ -1,0 +1,4 @@
+package config
+
+// DefaultFile is the name of the default wakatime config file.
+const DefaultFile = ".wakatime.cfg"
